requirements.txt
PyMuPDF>=1.23.0
gradio>=4.0.0
Pillow>=10.0.0
